Modification On Health bar (Edition01)

+++ Assets/Scripts/Core/Services/Damage/IDamageService.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageService
 {
-    void DealDamage(GameObject source, GameObject target, int amount);
+    void DealDamage(GameObject source, GameObject target, float amount);
 }
